requirements.txt
# llmaaj
pandas==2.0.3
pyarrow==14.0.1
python-dotenv==1.0.0
opentelemetry-api==1.33.0
opentelemetry-sdk==1.33.0
opentelemetry-exporter-otlp-proto-http==1.33.0
datasets==2.14.5
aiohttp
nest-asyncio
arize-phoenix==9.2.0

# tests
allure-pytest
pytest
pytest-assume
pytest-playwright
pytest-xdist
playwright
requests
streamlit
streamlit_autorefresh   
